feat(download): P4c-3-3 widen peer::Block to u64 (fix >4GiB narrowing)
P4c-3 spec §3.1 (M1). peer::Block.start/end were u32; decode_sending_part_i64
and decode_compressed_part_i64 did b.start = r.u64() into a u32 field, silently
truncating block offsets >= 2^32 (latent >4GiB corruption root cause).

Changes (mechanical widening, wire protocol untouched):
- c2c_messages.hpp: Block.start/end u32 -> u64. Legacy decode_sending_part/
  decode_compressed_part read r.u32() and zero-extend into the u64 field
  (legacy wire is u32). i64 variants now store u64->u64 (narrowing gone).
- part_file.{hpp,cpp}: PartFile::write_block start/end params u32 -> u64.
  Body was already u64-internal (explicit cast on line 56, u64 locals);
  seekp/seekg take 64-bit off-types, so >4GiB writes are correct.
- download.cpp: drop the two static_cast<u32>(b.start/b.end) narrowing casts
  at the write_block call sites (member path + peer_worker path).
- c2c_messages_test: add DecodeSendingPartI64Beyond4GiB asserting offsets
  0x100000000 and 0x100000000+184320 survive decode (would have failed pre-fix).

request_blocks (u32) / request_blocks_i64 (u64) dual wire paths retained;
caller selects by large_file flag (wire opcodes differ).

Build green; 214 tests, 209 passed, 5 skipped (no regression).

+++ include/ed2k/download/part_file.hpp
@@ -17,7 +17,7 @@ class PartFile {
   PartFile(const std::filesystem::path& path, std::uint64_t size, const FileHash& file_hash, std::vector<PartHash> part_hashes);
   bool open_for_write() const noexcept;
   std::vector<std::uint32_t> missing_parts_peer_has(const std::vector<bool>& peer_parts) const;
-  tl::expected<void,std::error_code> write_block(std::uint32_t start, std::uint32_t end, std::span<const std::byte> data);
+  tl::expected<void,std::error_code> write_block(std::uint64_t start, std::uint64_t end, std::span<const std::byte> data);
   bool is_block_done(std::size_t global_block) const noexcept;          // flat whole-file block index
   std::vector<std::size_t> pending_blocks() const;                      // flat global block indices
   bool complete() const noexcept;
